+++ src/main/java/com/ltonetwork/client/core/transacton/Anchor.java
@@ -6,7 +6,7 @@ import com.google.common.primitives.Longs;
 import com.ltonetwork.client.exceptions.BadMethodCallException;
 import com.ltonetwork.client.types.Encoding;
 import com.ltonetwork.client.utils.Encoder;
-import com.ltonetwork.client.utils.JsonObject;
+import com.ltonetwork.client.types.JsonObject;
 
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
