+++ src/main/java/com/ltonetwork/client/core/transacton/CancelLease.java
@@ -3,7 +3,7 @@ package com.ltonetwork.client.core.transacton;
 import com.google.common.primitives.Bytes;
 import com.google.common.primitives.Longs;
 import com.ltonetwork.client.exceptions.BadMethodCallException;
-import com.ltonetwork.client.utils.JsonObject;
+import com.ltonetwork.client.types.JsonObject;
 
 import java.nio.charset.StandardCharsets;
 
@@ -12,7 +12,6 @@ public class CancelLease extends Transaction {
     private final static byte TYPE = 9;
     private final static byte VERSION = 2;
     private final long leaseId;
-//    private final Lease lease;
 
     public CancelLease(int leaseId) {
         super(TYPE, VERSION, MINIMUM_FEE);
