+++ src/main/java/com/ltonetwork/client/core/transacton/MassTransfer.java
@@ -9,7 +9,7 @@ import com.ltonetwork.client.types.Address;
 import com.ltonetwork.client.types.Encoding;
 import com.ltonetwork.client.utils.CryptoUtil;
 import com.ltonetwork.client.utils.Encoder;
-import com.ltonetwork.client.utils.JsonObject;
+import com.ltonetwork.client.types.JsonObject;
 
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
