+++ src/main/java/com/ltonetwork/client/core/transacton/Sponsor.java
@@ -6,7 +6,7 @@ import com.ltonetwork.client.types.Address;
 import com.ltonetwork.client.exceptions.BadMethodCallException;
 import com.ltonetwork.client.exceptions.InvalidArgumentException;
 import com.ltonetwork.client.utils.Encoder;
-import com.ltonetwork.client.utils.JsonObject;
+import com.ltonetwork.client.types.JsonObject;
 
 import java.nio.charset.StandardCharsets;
