+++ src/main/java/com/ltonetwork/client/core/transacton/Transfer.java
@@ -8,7 +8,7 @@ import com.ltonetwork.client.exceptions.BadMethodCallException;
 import com.ltonetwork.client.exceptions.InvalidArgumentException;
 import com.ltonetwork.client.types.Encoding;
 import com.ltonetwork.client.utils.Encoder;
-import com.ltonetwork.client.utils.JsonObject;
+import com.ltonetwork.client.types.JsonObject;
 
 import java.nio.charset.StandardCharsets;
