+++ src/main/java/com/ltonetwork/client/types/Address.java
@@ -1,7 +1,6 @@
 package com.ltonetwork.client.types;
 
 import com.ltonetwork.client.exceptions.InvalidArgumentException;
-import com.ltonetwork.client.types.Encoding;
 import com.ltonetwork.client.utils.CryptoUtil;
 import com.ltonetwork.client.utils.Encoder;
